pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["z.ryo.ando <z.ryo.ando@sekaie.co.jp>"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.11.8"
fastapi = "^0.61.1"
sqlalchemy = "^1.3.19"
mysqlclient = "^2.0.1"
pyjwt = "^1.7.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
